Convert to using channel id's instead of channel names
This is necessary because YouTube has decided that channel names aren't
unique anymore. So this leaves two ways of requesting a channel:

  1. channel id
  2. username

The username is clearly the nicer one to use, because the channel id is
just a bunch of random characters. However, not every channel has a
username associated with it. I think this has to do with linking with G+
or something. So anyway, the only reliable, foolproof way of getting a
channel is through it's channel id.