Stop using the 'uploaded videos' playlist and use search instead
The 'uploaded videos' playlist just doesn't work well. It is slow to
update and regularly just does not contain certain videos. Instead let's
use a search. This search targets a videos for a specific channel id.

This seems to work better, doesn't miss videos and is always up to
date. On top of that it's much simpler because we just have to one
request instead of two. Also we don't need to keep a cache anymore. The
only issue with it is that some channels do not seem to set their
channelTitle. This means that those videos will show up as [] in the
HTML. There's no easy way to fix this since their channel name is not in
the response in any way. The only way to fix it would be to do a second
request.